EpiTxDb objects

The EpiTxDb class is a AnnotationDb type container for storing Epitranscriptomic information.

The information are typically stored on a per transcript and not as genomic coordinates, but the EpiTxDb class is agnostic to this. In case of genomic coordinates transcriptsBy will return modifications per chromosome.

# S4 method for EpiTxDb
organism(object)

# S4 method for EpiTxDb
seqinfo(x)

# S4 method for EpiTxDb
seqlevels(x)

# S4 method for EpiTxDb
as.list(x)

Arguments

x, object
a EpiTxDb object

Value

For

organism() and seqlevels() a character vector

seqinfo() a Seqinfo object

as.list() a list

See also

Examples

etdb_file <- system.file("extdata", "EpiTxDb.Hs.hg38.snoRNAdb.sqlite",
                        package="EpiTxDb")
etdb <- loadDb(etdb_file)
etdb

#> EpiTxDb object:
#> # Db type: EpiTxDb
#> # Supporting package: EpiTxDb
#> # Data source: snoRNAdb
#> # Organism: Homo sapiens
#> # Genome: hg38
#> # Coordinates: per Transcript
#> # Nb of modifications: 235
#> # Db created by: EpiTxDb package from Bioconductor
#> # Creation time: 2020-02-26 10:34:30 +0100 (Wed, 26 Feb 2020)
#> # EpiTxDb version at creation time: 0.99.0
#> # RSQLite version at creation time: 2.2.0
#> # DBSCHEMAVERSION: 1.0

# general methods
seqinfo(etdb) #

#> Seqinfo object with 9 sequences from hg38 genome; no seqlengths:
#>   seqnames  seqlengths isCircular genome
#>   NR_003285         NA         NA   hg38
#>   NR_003286         NA         NA   hg38
#>   NR_003287         NA         NA   hg38
#>   NR_004430         NA         NA   hg38
#>   NR_002716         NA         NA   hg38
#>   NR_003925         NA         NA   hg38
#>   NR_002756         NA         NA   hg38
#>   NR_004394         NA         NA   hg38
#>   NR_029422         NA         NA   hg38

seqlevels(etdb) # easy access to all transcript names

#> [1] "NR_003285" "NR_003286" "NR_003287" "NR_004430" "NR_002716" "NR_003925"
#> [7] "NR_002756" "NR_004394" "NR_029422"