Getting modification data from a EpiTxDb-object

modifications and modificationsBy are functions to extract modification annotation from a EpiTxDb object.

modifiedSeqsByTranscript returns a ModRNAStringSet from a EpiTxDb object and compatible RNAStringSet object. This used the combineIntoModstrings() function from the Modstrings package.

modifications(
  x,
  columns = c("mod_id", "mod_type", "mod_name"),
  filter = NULL,
  use.names = FALSE,
  ...
)

modificationsBy(
  x,
  by = c("seqnames", "mod_type", "reaction", "specifier", "specifier_type"),
  ...
)

modifiedSeqsByTranscript(x, sequences, ...)

# S4 method for EpiTxDb
modifications(
  x,
  columns = c("mod_id", "mod_type", "mod_name"),
  filter = NULL,
  use.names = FALSE
)

# S4 method for EpiTxDb
modificationsBy(
  x,
  by = c("seqnames", "modtype", "reaction", "specifier", "specifiertype")
)

# S4 method for EpiTxDb,DNAStringSet
modifiedSeqsByTranscript(x, sequences)

Arguments

x
a EpiTxDb

columns
Columns to include in the result. If the vector is named, those names are used for the corresponding column in the element metadata of the returned object. (default: columns = c("mod_id","mod_type","mod_name"))

filter
Either NULL or a named list of vectors to be used to restrict the output. Valid names for this list are: "mod_id", "mod_type", "mod_name", "sn_id", "sn_name", "rx_genename", "rx_ensembl", "rx_ensembltrans", "rx_entrezid", "spec_genename", "spec_type", "spec_ensembl", "spec_ensembltrans", "spec_entrezid" , "ref_type" and "ref". (default: filter = NULL)

use.names
TRUE or FALSE. If TRUE, the modification names are set as the names of the returned object. (default: use.names = FALSE)

...
Not used.

by
By which information type should the result be split into? A character value from one of the following values:

seqnames

mod_type

reaction

specifier

specifier_type

sequences
A RNAStringSet, which can be used as input for combineIntoModstrings(). See ?combineIntoModstrings for additional details.

Value

a GRanges object for modifications and a GRangesList for modificationsBy.

Examples

etdb_file <- system.file("extdata", "EpiTxDb.Hs.hg38.snoRNAdb.sqlite",
                        package="EpiTxDb")
etdb <- loadDb(etdb_file)
etdb

#> EpiTxDb object:
#> # Db type: EpiTxDb
#> # Supporting package: EpiTxDb
#> # Data source: snoRNAdb
#> # Organism: Homo sapiens
#> # Genome: hg38
#> # Coordinates: per Transcript
#> # Nb of modifications: 235
#> # Db created by: EpiTxDb package from Bioconductor
#> # Creation time: 2020-02-26 10:34:30 +0100 (Wed, 26 Feb 2020)
#> # EpiTxDb version at creation time: 0.99.0
#> # RSQLite version at creation time: 2.2.0
#> # DBSCHEMAVERSION: 1.0