Using the "select" interface on EpiTxDb objects

As expected for a AnnotationDb object, the general accessors select, keys, columns and keytypes can be used to get information from a EpiTxDb object.

# S4 method for EpiTxDb
select(x, keys, columns, keytype, ...)

# S4 method for EpiTxDb
columns(x)

# S4 method for EpiTxDb
keys(x, keytype, ...)

# S4 method for EpiTxDb
keytypes(x)

Arguments

x
a EpiTxDb object

keys, columns, keytype, ...
See AnnotationDb for more comprehensive description of the methods select, keys, columns and keytypes and their arguments.

Value

a data.frame object for select() and a character vecor for keytypes(), keys() and columns().

Examples

etdb_file <- system.file("extdata", "EpiTxDb.Hs.hg38.snoRNAdb.sqlite",
                         package="EpiTxDb")
etdb <- loadDb(etdb_file)
etdb

#> EpiTxDb object:
#> # Db type: EpiTxDb
#> # Supporting package: EpiTxDb
#> # Data source: snoRNAdb
#> # Organism: Homo sapiens
#> # Genome: hg38
#> # Coordinates: per Transcript
#> # Nb of modifications: 235
#> # Db created by: EpiTxDb package from Bioconductor
#> # Creation time: 2020-02-26 10:34:30 +0100 (Wed, 26 Feb 2020)
#> # EpiTxDb version at creation time: 0.99.0
#> # RSQLite version at creation time: 2.2.0
#> # DBSCHEMAVERSION: 1.0